make_bin_data
This function takes a data frame with or without a class and creates a one-hot encoded dataset, the parameters `comp` refers to if everything should be treated as a factor. If left `NULL` then any feature that has more that 5 levels(numerical as an example) will be placed into 5 buckets.